+++ src/OmniKinematics.h
@@ -13,24 +13,24 @@ public:
   OmniKinematics(uint8_t wheel) : wheelNumber(wheel), maxAllocateOutput(0.2){};
 
   /*
-    *   desc:   MDにかかる最大PWMを設定(0.1~1.0)
-    *   param:  maxPWM(float)
+    *   desc:   MDにかかる最大PWMを設定(10~95)
+    *   param:  maxPWM(int)
     *   return: none
     */
-  void setMaxPWM(float maxPwm)
+  void setMaxPWM(int maxPwm)
   {
     maxAllocateOutput = maxPwm;
   }
 
   /*
     *   desc:   ロボット全体の移動方向を指定し、MDにかけるPWMを取得
-    *   param:  x,y,yaw方向の移動量(float) 駆動輪数に応じたfloat配列
+    *   param:  x,y,yaw方向の移動量(int) 駆動輪数に応じたint配列
     *   return: none(引数に代入)
     */
-  void getOutput(float x, float y, float yaw, float pwm[]);
+  void getOutput(int x, int y, int yaw, int pwm[]);
 
 private:
   uint8_t wheelNumber;
-  float XVector, YVector, YawVector;
-  float maxAllocateOutput;
+  int XVector, YVector, YawVector;
+  int maxAllocateOutput;
 };
